:recycle: [filestorage] Move function CookiecutterFileStorage to test
